:recycle: rename test package
